improvements to common types
Renamed AVL tree references to AVL tree instances. This is more inline with
other uses of the name instance. Added functions to AVL trees that work
directly on AVL tree instances.

Fixed an uninitialized pointer bug int the queue. Added push and pop functions
to the linked-list for use as a stack.